Add files via upload
Modified schema -- course now has an ID, domain, and major requirement, removed latter two from the classes.js

Updated the courseform.html to reflect that
Added a classform.html and updatecourseform.html for pages to link to when editing a course / adding a class to enter in the related fields
Added a /showAll endpoint that shows all courses and classes together. Updated how they're shown to have links not just to delete but to add new courses/classes and edit the courses.